# feat: improve overall typescript performance