--- a/src/typescript-reporter/TypeScriptReporterOptions.ts
+++ b/src/typescript-reporter/TypeScriptReporterOptions.ts
@@ -15,6 +15,7 @@ type TypeScriptReporterOptions =
       extensions?: {
         vue?: TypeScriptVueExtensionOptions;
       };
+      profile?: boolean;
     };
 
 export { TypeScriptReporterOptions };