--- a/src/typescript-reporter/reporter/TypeScriptConfigurationParser.ts
+++ b/src/typescript-reporter/reporter/TypeScriptConfigurationParser.ts
@@ -23,7 +23,13 @@ function parseTypeScriptConfiguration(
     parsedConfigFile.errors.push(...customCompilerOptionsConvertResults.errors);
   }
 
-  return parsedConfigFile;
+  return {
+    ...parsedConfigFile,
+    options: {
+      ...parsedConfigFile.options,
+      configFilePath: configFileName,
+    },
+  };
 }
 
 export { parseTypeScriptConfiguration };