--- a/src/typescript-reporter/reporter/TypeScriptReporter.ts
+++ b/src/typescript-reporter/reporter/TypeScriptReporter.ts
@@ -12,6 +12,8 @@ import {
   createControlledTypeScriptSystem,
 } from './ControlledTypeScriptSystem';
 import { parseTypeScriptConfiguration } from './TypeScriptConfigurationParser';
+import { createPerformance } from '../../profile/Performance';
+import { connectTypeScriptPerformance } from '../profile/TypeScriptPerformance';
 
 function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration): Reporter {
   const extensions: TypeScriptExtension[] = [];
@@ -28,6 +30,7 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
   let solutionBuilder: ts.SolutionBuilder<ts.SemanticDiagnosticsBuilderProgram> | undefined;
 
   const diagnosticsPerProject = new Map<string, ts.Diagnostic[]>();
+  const performance = connectTypeScriptPerformance(createPerformance());
 
   if (configuration.extensions.vue.enabled) {
     extensions.push(createTypeScriptVueExtension(configuration.extensions.vue));
@@ -41,23 +44,50 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
     const diagnostics: ts.Diagnostic[] = [];
 
     if (configuration.diagnosticOptions.syntactic) {
+      performance.markStart('Syntactic Diagnostics');
       diagnostics.push(...builderProgram.getSyntacticDiagnostics());
-    }
-    if (configuration.diagnosticOptions.semantic) {
-      diagnostics.push(...builderProgram.getSemanticDiagnostics());
-    }
-    if (configuration.diagnosticOptions.declaration) {
-      diagnostics.push(...builderProgram.getDeclarationDiagnostics());
+      performance.markEnd('Syntactic Diagnostics');
     }
     if (configuration.diagnosticOptions.global) {
+      performance.markStart('Global Diagnostics');
       diagnostics.push(...builderProgram.getGlobalDiagnostics());
+      performance.markEnd('Global Diagnostics');
+    }
+    if (configuration.diagnosticOptions.semantic) {
+      performance.markStart('Semantic Diagnostics');
+      diagnostics.push(...builderProgram.getSemanticDiagnostics());
+      performance.markEnd('Semantic Diagnostics');
+    }
+    if (configuration.diagnosticOptions.declaration) {
+      performance.markStart('Declaration Diagnostics');
+      diagnostics.push(...builderProgram.getDeclarationDiagnostics());
+      performance.markEnd('Declaration Diagnostics');
     }
 
     return diagnostics;
   }
 
+  function emitTsBuildInfoFileForBuilderProgram(builderProgram: ts.BuilderProgram) {
+    if (
+      configuration.mode !== 'readonly' &&
+      parsedConfiguration &&
+      parsedConfiguration.options.incremental
+    ) {
+      const program = builderProgram.getProgram();
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      if (typeof (program as any).emitBuildInfo === 'function') {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (program as any).emitBuildInfo();
+      }
+    }
+  }
+
   return {
     getReport: async ({ changedFiles = [], deletedFiles = [] }: FilesChange) => {
+      if (configuration.profile) {
+        performance.enable();
+      }
+
       if (!system) {
         system = createControlledTypeScriptSystem(configuration.mode);
       }
@@ -83,6 +113,7 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
       if (!parsedConfiguration) {
         const parseConfigurationDiagnostics: ts.Diagnostic[] = [];
 
+        performance.markStart('Parse Configuration');
         parsedConfiguration = parseTypeScriptConfiguration(
           configuration.tsconfig,
           configuration.context,
@@ -94,6 +125,8 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
             },
           }
         );
+        performance.markEnd('Parse Configuration');
+
         if (parsedConfiguration.errors) {
           parseConfigurationDiagnostics.push(...parsedConfiguration.errors);
         }
@@ -123,6 +156,7 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
         // solution builder case
         // ensure watch solution builder host exists
         if (!watchSolutionBuilderHost) {
+          performance.markStart('Create Solution Builder Host');
           watchSolutionBuilderHost = createControlledWatchSolutionBuilderHost(
             parsedConfiguration,
             system,
@@ -137,25 +171,35 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
 
               // update diagnostics
               diagnosticsPerProject.set(projectName, diagnostics);
+
+              // emit .tsbuildinfo file if needed
+              emitTsBuildInfoFileForBuilderProgram(builderProgram);
             },
             extensions
           );
+          performance.markEnd('Create Solution Builder Host');
           solutionBuilder = undefined;
         }
 
         // ensure solution builder exists
         if (!solutionBuilder) {
+          performance.markStart('Create Solution Builder');
           solutionBuilder = ts.createSolutionBuilderWithWatch(
             watchSolutionBuilderHost,
             [configuration.tsconfig],
             {}
           );
+          performance.markEnd('Create Solution Builder');
+
+          performance.markStart('Build Solutions');
           solutionBuilder.build();
+          performance.markEnd('Build Solutions');
         }
       } else {
         // watch compiler case
         // ensure watch compiler host exists
         if (!watchCompilerHost) {
+          performance.markStart('Create Watch Compiler Host');
           watchCompilerHost = createControlledWatchCompilerHost(
             parsedConfiguration,
             system,
@@ -168,15 +212,21 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
 
               // update diagnostics
               diagnosticsPerProject.set(projectName, diagnostics);
+
+              // emit .tsbuildinfo file if needed
+              emitTsBuildInfoFileForBuilderProgram(builderProgram);
             },
             extensions
           );
+          performance.markEnd('Create Watch Compiler Host');
           watchProgram = undefined;
         }
 
         // ensure watch program exists
         if (!watchProgram) {
+          performance.markStart('Create Watch Program');
           watchProgram = ts.createWatchProgram(watchCompilerHost);
+          performance.markEnd('Create Watch Program');
         }
       }
 
@@ -192,7 +242,9 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
       });
 
       // wait for all queued events to be processed
+      performance.markStart('Queued Tasks');
       await system.waitForQueued();
+      performance.markEnd('Queued Tasks');
 
       // aggregate all diagnostics and map them to issues
       const diagnostics: ts.Diagnostic[] = [];
@@ -206,6 +258,11 @@ function createTypeScriptReporter(configuration: TypeScriptReporterConfiguration
           issues = extension.extendIssues(issues);
         }
       });
+
+      if (configuration.profile) {
+        performance.print();
+        performance.disable();
+      }
 
       return issues;
     },