--- a/test/e2e/TypeScriptConfigurationChange.spec.ts
+++ b/test/e2e/TypeScriptConfigurationChange.spec.ts
@@ -25,7 +25,7 @@ describe('TypeScript Configuration Change', () => {
   it.each([
     { async: true, webpack: '~4.0.0', typescript: '2.7.1', tsloader: '^5.0.0' },
     { async: false, webpack: '^4.0.0', typescript: '~3.0.0', tsloader: '^6.0.0' },
-    { async: true, webpack: '^5.0.0-beta.16', typescript: '~3.6.0', tsloader: '^7.0.0' },
+    { async: true, webpack: '^5.0.0-beta.16', typescript: '~3.7.0', tsloader: '^7.0.0' },
     { async: false, webpack: '^5.0.0-beta.16', typescript: '~3.8.0', tsloader: '^6.0.0' },
   ])(
     'change in the tsconfig.json affects compilation for %p',