--- a/test/unit/typescript-reporter/TypeScriptSupport.spec.ts
+++ b/test/unit/typescript-reporter/TypeScriptSupport.spec.ts
@@ -27,6 +27,7 @@ describe('typescript-reporter/TypeScriptSupport', () => {
         },
       },
       memoryLimit: 2048,
+      profile: false,
     };
   });