Aside from general performance improvements, I added a `profile` option to be able to measure timings and compare them with other tools, like tsc.

BREAKING CHANGE: 🧨 Changed default TypeScript compiler options to incremental: true and
the default mode to write-tsbuildinfo

**Related:**
https://github.com/vercel/next.js/pull/13529#issuecomment-635776458

**Results:**
I was testing it on my personal project. With incremental: true mode, the initial compilation time dropped from ~100s to ~30s (faster than initial webpack compilation). The incremental compilation time dropped from ~15s to ~2s.