@piotr-oles 2020-06-13
:tada: This PR is included in version 5.0.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/fork-ts-checker-webpack-plugin/v/5.0.0)
- [GitHub release](https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/releases/tag/v5.0.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: